backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-whois==0.8.0
apscheduler==3.10.4
smtplib2==0.2.1
twilio==8.10.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
plyer==2.1.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
databases[sqlite]==0.8.0
aiosqlite==0.19.0
jinja2==3.1.2 